Change library scene to abstract class
- Made Scene class in GeckoLibrary an interface with mostly abstract functions and much less member data;
- Moved implementation of NodeBasedScene to GeckoExample (it inherits from library Scene class and has all of the functionality that the Scene class had before)
- Moved functionality for rendering debug UI of node-based scenes to GeckoExample
- Moved functionality for loading a GLTF into a node-based scene to GeckoExample (this leaves the library without a model-loading class)
- Changed SceneManager API to better support derived scene classes
