+++ Example/Include/DebugSceneUIRenderer.h
@@ -1,0 +1,9 @@
+#pragma once
+#include "UI/DebugUIRenderer.h"
+
+class DebugSceneUIRenderer : public Gecko::DebugUIRenderer
+{
+public:
+	// Draw debug info of node-based scenes
+	static bool RenderDebugUI(Gecko::ApplicationContext& ctx);
+};
