+++ Example/Include/GLTFSceneLoader.h
@@ -1,0 +1,18 @@
+#pragma once
+
+#include "Defines.h"
+
+namespace Gecko
+{
+	class Scene;
+	class ApplicationContext;
+	using SceneHandle = u32;
+}
+
+class GLTFSceneLoader
+{
+public:
+	// Loads a GLTF scene into a NodeBasedScene object and puts it in SceneManager (which lives in passed context)
+	// Returns a SceneHandle generated by the SceneManager, which can be used to access the scene later
+	[[nodiscard]] static Gecko::SceneHandle LoadScene(const std::string& pathString, Gecko::ApplicationContext& ctx);
+};
