--- Include/Rendering/Frontend/Scene/GLTFSceneLoader.h
@@ -1,17 +1,0 @@
-#pragma once
-
-#include "Defines.h"
-
-namespace Gecko
-{
-
-	class Scene;
-	class ApplicationContext;
-
-	class GLTFSceneLoader
-	{
-	public:
-		[[nodiscard]] static u32 LoadScene(const std::string& pathString, ApplicationContext& ctx);
-	};
-
-}
